Set quad_current to contain partial current; pass into LOO.
As a very first step, quad_current now stores partial currents,
and are passed into LOO. LOO prints out both quad_current and
current and the two agree (of course....). The next step is to
make quad_current to store actual quad current, and verify that
the two quad currents in the same half-space of a surface sums
up to be the corresponding partial current.