Fixed a bug in tallying quad current; should be working for 2D prob.
After implementing the previous commit, partial currents on reflective
boundaries are non-zero and equal in value for the same surface (which
is what we expect). However that was not the case for quad currents.

As it turned out, there was a loop over the three directions where we
treat the case where the closest mesh surface is in that
direction. For 2D problems, even if we don't care about tallying on
the z-surfaces, we cannot omit entering the routine where z-planes are
the closet surfaces to the neutron in the direction that the neutron
is travelling, because the routine also updates the coordinates of the
neutron and the indexes of the mesh it is in as the routine
artificially moves the neutron from one surface to another to make
sure all the surfaces in the path of this event is tallied. So current
treatment just makes sure that j = 3 loop is still entered, it simply
does not tallies currents on the z-planes.