Updated track generation routine to store both x, y indexs of cell
In OpenMOC's LOO implementation, cell index is flattened to be one
term to keep the LOO array types with less number of dimensions. In
this implementation, more proper data structures like meshElement and
surfaceElement are implemented with consistant indexing into data
structure that it makes more sense for the track generation to produce
cell indexes in both x and y. generate2dTrack() is updated accordingly
to generate both x and y indexes for cells now.