Processed volume and area out of _scalar_flux, _quad_current.
This commit realized that the fluxes and currents passed in from
openmc's cmfd object are the volume-integrated and area-integrated
terms, respectively. This commmit divided the volume and surface area
terms out of _scalar_flux and _quad_current, so that they are what we
commonly define them as.

As part of this commit, now _volume stores the actual volume of each
mesh cell, and _area stores the x-, y- and z-direction surface area of
each mesh cell.