Updated the k passing from openmc into LOO
I used to pass in the cmfd % keff_bal term. Though it is unclear to me
what that term actually means because it looks like currently no other
routine actually uses this term.

This commit uses the keff term defined in the global file and shared
among all files as the keff generated from MC and passed it into LOO.