Fixed a bug in track index convention.
When I first implemented computeQuadSrc() its track index convention
was incorrect. This has been fixed in this commit. The current track
index convention is illustrated in front of the generate2dTrack()
routine. Also the getSurfaceLength(..)  routine has been updated to
getSurfaceArea(..) routine which returns the starting or ending
surface area for a given track.