Removed the sine of the one polar angle from Loo routine.
Duringresearch meeting today, we discussed that the polar quadrature
set (like the Tabuchi-Yamamoto set that we use for OpenMOC) is just a
way of handling the third dimension for deterministic method.

In openmc, when we start accelerating the real 3D problem, there is no
reason to worry about using polar angle quadrature set anymore. Even
for the 2D Loo problem we are solving right now, we essentially are
just performing acceleration in pure vanilla 2D, with no
representation of the third dimension.