Updated Loo routines: fission source, normalization, compute eps
The flow of the executeLoo() concerning the normalization and such
looks like the following: first computes and store the average of
initial fission source (this is entirely generated by openmc), and
save a copy of this into fission_source (which will lag
_fission_source) by one iteration. Then at the end of each Loo sweep,
we call normalization(old_src) to compute the new fission source, and
normalize terms such that current fs' average remains old_src. Then we
compute the L2 norm between the current and old fs. Then we do
book-keeping to store _fission_source into fission_source, before
moving on to the next iteration.

* computeFissionSource(old_avg) does a couple of things, including
  computing the most-up-to-date FS and store in _fission_source,
  computes and store (in _rms) this distribution's deviation from
  flat, and return normalization factor that scales the avg of this
  distribution to be old_avg.

* normalization(old_avg) normalizes internal fission source, scalar
  fluxes, quad flux and leakage such that the average fission source
  is old_avg.

* computeL2Norm(old_fs) computes the L2 norm relative difference
  between the old fission source old_fs and the current fission source
  _fission_source.