Lag fission source alone, as opposed to the total source.
The first round of results from the previous implementation
revealed that LOO fluctuates more than CMFD. While it is possible
there is a bug that has not been found in the Loo routine, one
thing worth testing is to lag the fission source alone as opposed
to the total source.

The idea is, in methods of characteristics, one transport sweep is
essentially putting the lagged total source on the RHS of the
transport equation and solve for updated scalar fluxes, which is
why we needed the lagged total source in LOO to be consistant with
the high order.

One possibility is that, in Monte Carlo, one sweep is essentially
using the lagged fission source and solve for updated reaction
rates. Hence to be consistant with MC, LOO should lag the fission
source as well (and keep the scattering source up to date).