Fixed and updated commit 39e9b81f on printing to log file.
Issue with commit 39e9b81f9e2be58b97904c7cab0a4ce1c4ec16d8 is that
it was opening log file with the "w" flag which overwrite the
messages from the previous batch.

This commit opens log file with "a" (append) option so there is no
over-written. Also the pointer to the file is now an internal
variable so that Loo object all have access to it.

To-do: log file name is currently just log.txt which means that if
multiple ran was made they would be appended to the same file.
This commit contains routine for getting the current time, though
the issue is that the Loo object is constructed for every batch of
MC, so passing the time into log file name would result in one
unique log file for every batch of MC. The only clean way to involve
some additional code in Fortran because the C++ code does not know
whether we can in the same run or not.