Fixed commit 7c64b0081 on lagging fission source.
Issue with commit 7c64b0081c922e12c20585677a0e7ea686041af9: fission
source passed in from openmc does not include k and was normalized
such that the sum adds up to be 1.0, whereas the fission source
computed internally in the computeQuadSource(..) routine was already
divided by k and also it was normalized differently.

This commit fixed the above problems:

1) fission source computed internally in computeQuadSource(..) does
not include k anymore to be consistant with the _old_source passed
in from openmc;

2) the _old_source passed in from openmc (of iteration m, i.e., the
fission source that generated this batch of MC) is normalized before
LOO sweeps such that its average is the same as the initial fission
source (of iteration m+1/2, i.e., the fission source generated by this
batch of MC).