Implemented track generation for nx1 meshes.
In the special case of n by 1 meshes, it looks like we can walk
through all the tracks in one loop. Track generations are updated
to support the nx1 case, and laydown were printed out to make sure
that it works for both even number of cells and odd number of cells.