Added loo_run flag set in settings.xml, compute cmfd % loo_entropy
loo_run logical is set to be true if <run_loo> is set to be true in
settings.xml. This logic is not quite mature yet -- currently to run
loo we need both <run_cmfd> and <run_loo> to be true: It would run
both accelerations, and the <feedback> tag in cmfd.xml would control
whether to feedback the loo-generated fission source or not.

We now also declare cmfd % loo_entropy, which stored the shannon
entropy calculated from the loo-generated fission source. This is
printed to screen when loo is on.