Fixed geometry's tally currents routine from merging with upstream
I changed the neutron maneuvering mechanism in the score neutrons
routine in geometry.F90 to make it work with LOO (it was working
before by cancellation for CMFD). Though those changes appeared
to be overwritten during merging with the upstream develop branch.

I need to keep this in mind moving forward to make sure geometry
routine stays relevant for LOO with future merging...