Temporary fix nx1 with vac BC in x: albedos, logics for vac;
This commit is not a general one but specific for a nx1 problem
with vac BCs in x-direction and reflective BCs in y-direction:

1) _albedo[s] passed in from OpenMC turned out to be all 1.0 even
when two surfaces are vaccum. This needs to be looked into for
the fugure, but for now I force the two BCs in x-direction to be
0.0.

2) logics for determing whethe a given track leaves a vac BC needs
to be reconsidered. Luckly in this case the two boundaries in the
y-direction are reflective; the current logics does not work if
the y-direction BCs are vacuum.