Modification of slab problem: 101x1x1, turn loo on, no feedback.
Changed the number of mesh cells from even (100) to odd (101) so
that the middle cell has the peak value (cell 51). Also I should
have checked in the addition of loo flag earlier when the feature
was first implemented. This current setting file turns on both
CMFD and LOO but does not feedback into openmc.