Passed back the correct LOO k.
Previously what I considered as the LOO calculated k turned out
to be just a copy of the MC keff (which was the initial k passed
into LOO to start the LOO calculation with). This has been fixed,
now we are actually passing the LOO calculated keff back into
openmc for printing to screen and such.