Use source_bank as old fission source in LOO.
Playing around with what's appropriate for LOO's old FS. The
current though is to use the source_bank generated at the end of
each batch as the old FS for the next batch.

My current understanding is that accelerations should be blind
to the fission_bank, because after MC generates the fission_bank,
it then proceeds to sample from fission_bank to produce
source_bank during finalize_generation; then in finalize_batch
accelerations are performed, and reweight will be done on
source_bank as well. Hence technically fission_bank never comes in
play for accelerations.