Modifications to suppport MPI operations in computating source bank
This commits need to be double-checked -- essentially before this
commit, the serial version of the code runs, but not when mpi is
turned on. I think the issue is with coping entropy_s into
entropy_s_old, so I put a "if master" statement around it, and I
think by now the entropy_s information has been collected on the
master node, though I really need to double check on this.