Improve the initial condition of LOO when LOO is on from batch 1.
Background: previously I didn't handle the case when acceleration is
turned on from the very first batch. That is, during the first batch
of MC, the old FS and old k passed into LOO used to be ill-defined
because the corresponding parameters from Fortran may or may not be
initialized).

Action: this commit checks for the k and old FS passed into LOO when LOO
starts, and if they appears to be ill-defined, k is set to be 1.0, and
old FS is set to be flat.

Effect: LOO does not have ill-defined parameters (e.g., 0.0 or nan) when
starting.

To-do: look into why after the first swip LOO generates 0.0 FS and nan
k when acceleration is turned on from the very beginning. For now this
is not a big deal because we don't care that much about the results of
the first batch that much.