Fixed bug: the quad currents passed into LOO should be normalized.
Background: I've noticed for a while that LOO is a lot more sensitive
to the tally flushing pattern than CMFD.

Issue: it turned out the quad currents were not normalized to remove
the effect from increasing number of batches, hence they grow as
number of batches accumulates. At the same time, this is not an issue
for the cross-sections because they are computed by reaction rates
tallies divided by flux tallies hence the effect of increasing
histories do not show up. Similarly for CMFD, the absolute values of
the currents were not used (only the ratios of the net currents to
fluxes were used), hence CMFD was affected by the growing histories
neither.

Edit: in this commit I added normalization of the quad currents
(by the same amount of the scalar fluxes etc as part of the
normalizeTallies() routine), then call compute quad flux after
quad currents are normalized.

To do: so far quad currents (hence quad fluxes) are normalized at the
initialization stage of LOO only. I still need to re-think whether
quad fluxes need to be re-normalized during each LOO sweep or not. An
earlier commit (29c2928) specifically noted that quad fluxes cannot be
re-normalized during every LOO sweep, as it was breaking the test
problem back then.