Implemented and disabled LOO feature to read reference parameters.
This is a functionality that's turned on by the pound defined
reference key word. This feature should no long be that necessary
now that we can use the read in reference parameter function
defined in CMFD to read in parameters once and apply onto both
methods.

Also a couple of pound defined values I am playing around have been
moved from the loo.h to loo.cpp, and P0 (the 1-polar angle
Tabuchi-Yamamoto quadrature set) has been added to LOO for now (though
I believe with or without P0 should not change the answer of LOO at
this point.