Implemented "cosine" option in setting.xml for initial source
That is, essentially based on the "fission" option, except that
instead of equally probable from every place, it samples from a
distribution s.t. that sampled initial source is a chopped cosine
using the extrapolation distance specified by user in the
setting.xml files source tag's last entry in the parameters field.