Fix bug from last commit: openmc_src_rate need to be reset.
Almost all of the xxx_rate tallies added into cmfd_data.F90 in the
last commit are being over-written at each generation which was why
they didn't have to be explicitly reset. Except openmc_src_rate is
being accumulated based on the energy levels, which is why it needs to
be reset at each generation.