Added option to accumulate all tallies after certain batches.
The default functionality has not changed in that commit; this commit
simply offers options of accumulating all tallies if turned on. That
is, instead of over-written the corresponding positions in the tally
arrays, we now accumulate the tally arrays, and in keeping the default
behavior I reset the corresponding positions at the beginning of each
batch for now.