Intial implementation for expanding moving window, w/ <roll_up_to>
Before: a fixed rolling window defined by <n_save> in cmfd.xml.

Now: a <roll_up_to> in cmfd.xml would specify the maximum rolling
window size allowed (for now it should be the largest power of 2 less
than that of the total number of accelerated batches; see fixme
below). Then the rolling window size is expended by a factor of 2 each
time we hit the batch of number power of 2.

For example, the starting sequence and their corresponding saved
previous batch numbers are (in the order they are stored in the
history array):

 1: 1
 2: 1, 2
 3: 3, 2
 4: 2, 3, 4, -
 5: 2, 3, 4, 5
 6: 6, 3, 4, 5
 7: 6, 7, 4, 5
 8: 4, 5, 6, 7, 8, -, -, -
 9: 4, 5, 6, 7, 8, 9, -, -
10: 4, 5, 6, 7, 8, 9,10, -
11: 4, 5, 6, 7, 8, 9,10,11
12:12, 5, 6, 7, 8, 9,10,11
13:12,13, 6, 7, 8, 9,10,11
14:12,13,14, 7, 8, 9,10,11
15:12,13,14,15, 8, 9,10,11
16: 8, 9,10,11,12,13,14,15,16, -, -, -, -, -, -, -

FIXME: 1) do I need an update on cmfd_n_save in the case of expanding
rolling window to ensure that it is in fact a power of two? or it does
not have to be? 2) what about not start acceleration right away? 3)
this initial implementation is not very memory efficient.