Saved two more terms to state point to restart rolling parameters
With this and the previous two commits, we should be able to restart
now and reproduce all the solutions (MC FS, CMFD solution, LOO
solution). I did two simple test one is the LOO-accelerated case
without calling the <roll_up_to>, and the other is with
<roll_up_to>. Both case the restarted cases were able to reproduce the
screen output of the original cases.