Improve unaccelerated MC: entropy_p print to same location in fs.dat
This commit concerns the true unaccelerated MC (both run_cmfd and
run_loo false) to make sure that entropy_p is still printed at the 9th
location to be consistent with the accelerated case such that it is
easier on the python script.

TODO: obviously this is not very clean... should clean up this feature
later.