Fix commit f80af354: missed a return statement; add helping messages
commit f80af354 missed a return statement after the added
handle_lost_neutron() routine, which caused the neutron whose location
is in an undefined universe to continue going, generating a
segmentation fault from accessing an undefined universe. This commit
added the needed return statement following the commits from the
upstream repo (see commit f80af354 for details). Also some additional
info are added so that when things fail there are more information.

TODO: I understand that the upstream repo added the
handle_lost_particle here as there are not that many neutrons that get
into these situations; though in general I think it's a doable thing
to figure out why when neutrons arrive on a surface it has trouble
figure out where it is.