Improve performance by tally either current or quad_current
Previously I was always tallying both partial currents and quad
currents regardless of which acceleration method is on. Back then I
wasn't worried about performance at all so it's easy to just have both
on so that I have more information and do more debuggin etc.

This commit adds condition for tallying currents, such that partial
currents are only tallied and computed and stored for CMFD, and quad
currents are only tallied and computed and stored for LOO.

Outcome: this commit should reduce CPU time since the amount of
current tallying is reduced by roughly half. Though when tested on a
real server (eofe4.mit.edu in this case), the run time differs from
one node to another, so it's hard to actually observe whether the run
time is reduced or not. Though I think logically it is pretty obvious
that this commit should save time, I just don't know how much it does
when running in parallel.