Edits to get LOO working with core map defined in cmfd.xml
Previously LOO only works with regular-shaped grids (for instance, nx1
and nxn) where low-order tracks cover all the meshes.

This commit allows LOO to parse in the core map defined in cmfd.xml
and only lay down tracks in the active meshes. This appears to be
necessary to get LOO to converge to the right answer (i.e, when we
allow mesh of zero reaction to be swept through, they mess up the
total solution).

TODO: need to make sure the implementation is still general that
problem without core map still works.